+++ tests/Feature/RouteTests/CreateTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Feature\RoutTests;
+namespace Tests\Feature\RouteTests;
 
 use App\Models\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
@@ -14,14 +14,14 @@ class CreateTest extends TestCase
      *
      * @return void
      */
-    public function test_acount_profile_controller_create()
+    public function test_account_profile_controller_create()
     {
         $user = User::factory()->create();
         $_GET = ['profile' => 14];
         $response = $this->actingAs($user)->get(route('account.profiles.create', $_GET));
         $response->assertOk();
     }
-    public function test_acount_users_controller_create()
+    public function test_account_users_controller_create()
     {
         $user = User::factory()->create();
         $user->role_id = 1;
